Renamed files in `smelter-api` to `moq_server`
