+++ smelter-api/src/input.rs
@@ -2,8 +2,8 @@ mod decklink;
 mod decklink_into;
 mod hls;
 mod hls_into;
-mod moq;
-mod moq_into;
+mod moq_server;
+mod moq_server_into;
 mod mp4;
 mod mp4_into;
 mod rtmp;
@@ -22,7 +22,7 @@ mod side_channel;
 
 pub use decklink::*;
 pub use hls::*;
-pub use moq::*;
+pub use moq_server::*;
 pub use mp4::*;
 pub use rtmp::*;
 pub use rtp::*;
